Added barebones print_expr function and fixed power rule bug
I added a function that will print expressions but haven't yet added code for most cases. However, the unfinished cases will be easy to fill in since they are just small variations on the code for already-finished cases. I also fixed a bug with how the power rule was implemented, noting that (f^a)'=a*f'*f^{a-1} and not (f^a)'=af^{a-1} since chain rule is needed if f has a derivative not equal to 1. Made bare minimum changes needed for my first regression test to work (see slurpslurp_regression), lots of missing cases I need to fill in on match statements, but that is all easy as mentioned before.